Optimized level loading: pre-search for gamestate with scene
When a level loads, it now checks to see if any gamestate has this level
in the open scenes list. If not, it doesn't load any objects. If so, it
starts each of its object's searches at the most recent game state where
the scene was open instead of the current game state