Reorganized/ removed PNG files from the repository
Because image assets go in a separate private online storage space, and not in the public GitHub repository, the PNGs have been removed from this repository and the repository now ignores .PNG files.

NOTE: .PNG.META files are still in the repository because Unity needs them to render the images correctly.