Possibly reduced the splash image delay
There's a bit of a delay after the unity logo disappears before the splash screen appears. I think it's due to having to load the player scene and the other scenes.
Beyond this measure, the only other way to make it faster is to make the splash screen its own scene